requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.10
Requests==2.32.3
SpeechRecognition==3.14.1
